+++ scripts/subgenome_ref_files_3.sh
@@ -1231,27 +1231,33 @@ main () {
 
 	touch ${projdir}/compress_done.txt
 
+	if [[ "$samples_list" == "samples_list_node_1.txt" ]]; then
+		mv ${projdir}/preprocess/processed/${i%.f*}_*_precall.bam* ${projdir}/preprocess/ 2> /dev/null
+	fi
+	if [[ $nodes -gt 1 ]] && [[ "$samples_list" == "samples_list_node_1.txt" ]]; then
+		mv ${projdir}/preprocess/${i%.f*}_*_precall.bam* ${projdir}/preprocess/processed/ 2> /dev/null
+		touch ${projdir}/call0
+	fi
 
+	while [[ test -f "${projdir}/call0" ]]; do sleep 30; done
 
-	if [ "$alignments" == 0 ]; then
-		if [[ $nodes -gt 1 ]]; then
-			mkdir -p /tmp/${samples_list%.txt}/refgenomes /tmp/${samples_list%.txt}/samples /tmp/${samples_list%.txt}/preprocess /tmp/${samples_list%.txt}/snpcall
+	if [[ $nodes -gt 1 ]] && [[ test -f "${projdir}/call0" ]]; then
+		mkdir -p /tmp/${samples_list%.txt}/refgenomes /tmp/${samples_list%.txt}/samples /tmp/${samples_list%.txt}/preprocess /tmp/${samples_list%.txt}/snpcall
+		touch ${projdir}/queue_move_${samples_list%.txt}
+		queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
+		while [[ "$queue_move" -gt 1 ]]; do
+			rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
 			touch ${projdir}/queue_move_${samples_list%.txt}
 			queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-			while [[ "$queue_move" -gt 1 ]]; do
-				rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-				touch ${projdir}/queue_move_${samples_list%.txt}
-				queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
+		done
+		cp -rn ${projdir}/refgenomes/* /tmp/${samples_list%.txt}/refgenomes/ &&
+		if [[ "$lib_type" == "RRS" ]]; then
+			for i in $(cat ${projdir}/${samples_list} ); do
+				cp ${projdir}/preprocess/processed/${i%.f*}_*_precall.bam* /tmp/${samples_list%.txt}/preprocess/ 2> /dev/null &&
+				wait
 			done
-			cp -r ${projdir}/refgenomes/* /tmp/${samples_list%.txt}/refgenomes/ &&
-			if [[ "$lib_type" == "RRS" ]]; then
-				for i in $(cat ${projdir}/${samples_list} ); do
-					cp ${projdir}/preprocess/${i%.f*}_*_precall.bam* /tmp/${samples_list%.txt}/preprocess/ 2> /dev/null &&
-					wait
-				done
-			fi
-			rm ${projdir}/queue_move_${samples_list%.txt}
 		fi
+		rm ${projdir}/queue_move_${samples_list%.txt}
 	fi
 
 
@@ -1259,19 +1265,6 @@ main () {
 	cd preprocess
 	mkdir -p processed
 
-	if [[ ! -f "${projdir}/call0_${samples_list}" ]]; then touch "${projdir}/call0_${samples_list}"; fi
-	if [[ "$samples_list" == "samples_list_node_1.txt" ]]; then
-		align=$(ls ${projdir}/call0_samples_list_node_* | wc -l)
-		while [[ "$align" -lt $nodes ]]; do sleep 300; align=$(ls ${projdir}/call0_samples_list_node_* | wc -l); done
-		if [[ $align == $nodes ]]; then
-			rm ${projdir}/call0_${samples_list}
-			if [[ -z "$(ls -A ./processed/ &> /dev/null)" ]]; then
-				:
-			else
-				mv $projdir/preprocess/processed/*_precall.bam* $projdir/preprocess/
-			fi
-		fi
-	fi
 
 	if [[ "$samples_list" == "samples_list_node_1.txt" ]]; then
 		if [[ "$joint_calling" == true ]]; then
@@ -1603,7 +1596,9 @@ main () {
     			fi
     			mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}.hold.g.vcf.gz" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}.g.vcf.gz" && \
     			mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}.hold.g.vcf.gz.tbi" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}.g.vcf.gz.tbi" &&
-    			wait
+    			mv ${i%.f*}_${ref1%.f*}_${ref2%.f*}_precall.bam* ./processed/
+					mv ${i%.f*}_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}_precall.bam* ./processed/
+					wait
     		fi
     	fi ) &
     	if [[ $(jobs -r -p | wc -l) -ge $gN ]]; then
@@ -1611,20 +1606,6 @@ main () {
     	fi
     done
     wait
-		if [[ $nodes -gt 1 ]] && test -f ${projdir}/GBSapp_run_node_1.sh; then
-			touch ${projdir}/queue_move_${samples_list%.txt}
-			queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-			while [[ "$queue_move" -gt 1 ]]; do
-				rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-				touch ${projdir}/queue_move_${samples_list%.txt}
-				queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-			done
-			mv *_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}_precall* ./processed/ &&
-			rm ${projdir}/queue_move_${samples_list%.txt}
-		else
-			mv *_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}_precall* ./processed/ &&
-			wait
-		fi
 		touch "${projdir}/call123_${samples_list}"
 
     if [[ "$samples_list" == "samples_list_node_1.txt" ]] && test ! -f ${projdir}/snpcall/${pop}_${ref1%.f*}_${ref2%.f*}_${ref3%.f*}_${ploidy}x_raw.vcf*; then
@@ -1747,7 +1728,8 @@ main () {
     			fi
     			mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}.hold.g.vcf.gz" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}.g.vcf.gz" && \
     			mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}.hold.g.vcf.gz.tbi" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref2%.f*}.g.vcf.gz.tbi" &&
-    			wait
+    			mv ${i%.f*}_${ref1%.f*}_${ref2%.f*}_precall.bam* ./processed/
+					wait
     		fi
     	fi ) &
     	if [[ $(jobs -r -p | wc -l) -ge $gN ]]; then
@@ -1755,20 +1737,6 @@ main () {
     	fi
     done
     wait
-		if [[ $nodes -gt 1 ]] && test -f ${projdir}/GBSapp_run_node_1.sh; then
-		  touch ${projdir}/queue_move_${samples_list%.txt}
-		  queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-		  while [[ "$queue_move" -gt 1 ]]; do
-		    rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-		    touch ${projdir}/queue_move_${samples_list%.txt}
-		    queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-		  done
-		  mv *_${ref1%.f*}_${ref2%.f*}_precall* ./processed/ &&
-		  rm ${projdir}/queue_move_${samples_list%.txt}
-		else
-		  mv *_${ref1%.f*}_${ref2%.f*}_precall* ./processed/ &&
-		  wait
-		fi
 		touch "${projdir}/call12_${samples_list}"
 
     if [[ "$samples_list" == "samples_list_node_1.txt" ]] && test ! -f ${projdir}/snpcall/${pop}_${ref1%.f*}_${ref2%.f*}_${calcploidy}x_raw.vcf*; then
@@ -1891,7 +1859,8 @@ main () {
     			fi
     			mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref3%.f*}.hold.g.vcf.gz" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref3%.f*}.g.vcf.gz" && \
     			mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref3%.f*}.hold.g.vcf.gz.tbi" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}_${ref3%.f*}.g.vcf.gz.tbi" &&
-    			wait
+    			mv ${i%.f*}_${ref1%.f*}_${ref3%.f*}_precall.bam* ./processed/
+					wait
     		fi
     	fi ) &
     	if [[ $(jobs -r -p | wc -l) -ge $gN ]]; then
@@ -1899,20 +1868,6 @@ main () {
     	fi
     done
     wait
-		if [[ $nodes -gt 1 ]] && test -f ${projdir}/GBSapp_run_node_1.sh; then
-			touch ${projdir}/queue_move_${samples_list%.txt}
-			queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-			while [[ "$queue_move" -gt 1 ]]; do
-				rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-				touch ${projdir}/queue_move_${samples_list%.txt}
-				queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-			done
-			mv *_${ref1%.f*}_${ref3%.f*}_precall* ./processed/ &&
-			rm ${projdir}/queue_move_${samples_list%.txt}
-		else
-			mv *_${ref1%.f*}_${ref3%.f*}_precall* ./processed/ &&
-			wait
-		fi
 		touch "${projdir}/call13_${samples_list}"
 
     if [[ "$samples_list" == "samples_list_node_1.txt" ]] && test ! -f ${projdir}/snpcall/${pop}_${ref1%.f*}_${ref3%.f*}_${calcploidy}x_raw.vcf*; then
@@ -2035,7 +1990,8 @@ main () {
     			fi
     			mv "${projdir}/snpcall/${i%.f*}_${ref2%.f*}_${ref3%.f*}.hold.g.vcf.gz" "${projdir}/snpcall/${i%.f*}_${ref2%.f*}_${ref3%.f*}.g.vcf.gz" && \
     			mv "${projdir}/snpcall/${i%.f*}_${ref2%.f*}_${ref3%.f*}.hold.g.vcf.gz.tbi" "${projdir}/snpcall/${i%.f*}_${ref2%.f*}_${ref3%.f*}.g.vcf.gz.tbi" &&
-    			wait
+    			mv ${i%.f*}_${ref2%.f*}_${ref3%.f*}_precall.bam* ./processed/
+					wait
     		fi
     	fi ) &
     	if [[ $(jobs -r -p | wc -l) -ge $gN ]]; then
@@ -2043,20 +1999,6 @@ main () {
     	fi
     done
     wait
-		if [[ $nodes -gt 1 ]] && test -f ${projdir}/GBSapp_run_node_1.sh; then
-			touch ${projdir}/queue_move_${samples_list%.txt}
-			queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-			while [[ "$queue_move" -gt 1 ]]; do
-				rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-				touch ${projdir}/queue_move_${samples_list%.txt}
-				queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-			done
-			mv *_${ref2%.f*}_${ref3%.f*}_precall* ./processed/ &&
-			rm ${projdir}/queue_move_${samples_list%.txt}
-		else
-			mv *_${ref2%.f*}_${ref3%.f*}_precall* ./processed/ &&
-			wait
-		fi
 		touch "${projdir}/call23_${samples_list}"
 
     if [[ "$samples_list" == "samples_list_node_1.txt" ]] && test ! -f ${projdir}/snpcall/${pop}_${ref2%.f*}_${ref3%.f*}_${calcploidy}x_raw.vcf*; then
@@ -2178,7 +2120,9 @@ main () {
 						fi
 						mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}.hold.g.vcf.gz" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}.g.vcf.gz" && \
 						mv "${projdir}/snpcall/${i%.f*}_${ref1%.f*}.hold.g.vcf.gz.tbi" "${projdir}/snpcall/${i%.f*}_${ref1%.f*}.g.vcf.gz.tbi" &&
+						mv ${i%.f*}_${ref1%.f*}_precall.bam* ./processed/
 						wait
+
 					fi
 				fi ) &
 				if [[ $(jobs -r -p | wc -l) -ge $gN ]]; then
@@ -2186,20 +2130,6 @@ main () {
 				fi
 			done
 			wait
-			if [[ $nodes -gt 1 ]] && test -f ${projdir}/GBSapp_run_node_1.sh; then
-				touch ${projdir}/queue_move_${samples_list%.txt}
-				queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-				while [[ "$queue_move" -gt 1 ]]; do
-					rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-					touch ${projdir}/queue_move_${samples_list%.txt}
-					queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-				done
-				mv *_${ref1%.f*}_precall* ./processed/ &&
-				rm ${projdir}/queue_move_${samples_list%.txt}
-			else
-				mv *_${ref1%.f*}_precall* ./processed/ &&
-				wait
-			fi
 			touch "${projdir}/call1_${samples_list}"
 
 			if [[ "$samples_list" == "samples_list_node_1.txt" ]] && test ! -f $${projdir}/snpcall/{pop}_${ref1%.f*}_${ploidy_ref1}x_raw.vcf*; then
@@ -2321,6 +2251,7 @@ main () {
 						fi
 						mv "${projdir}/snpcall/${i%.f*}_${ref2%.f*}.hold.g.vcf.gz" "${projdir}/snpcall/${i%.f*}_${ref2%.f*}.g.vcf.gz" && \
 						mv "${projdir}/snpcall/${i%.f*}_${ref2%.f*}.hold.g.vcf.gz.tbi" "${projdir}/snpcall/${i%.f*}_${ref2%.f*}.g.vcf.gz.tbi" &&
+						mv ${i%.f*}_${ref2%.f*}_precall.bam* ./processed/
 						wait
 					fi
 				fi ) &
@@ -2329,20 +2260,6 @@ main () {
 				fi
 			done
 			wait
-			if [[ $nodes -gt 1 ]] && test -f ${projdir}/GBSapp_run_node_1.sh; then
-				touch ${projdir}/queue_move_${samples_list%.txt}
-				queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-				while [[ "$queue_move" -gt 1 ]]; do
-					rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-					touch ${projdir}/queue_move_${samples_list%.txt}
-					queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-				done
-				mv *_${ref2%.f*}_precall* ./processed/ &&
-				rm ${projdir}/queue_move_${samples_list%.txt}
-			else
-				mv *_${ref2%.f*}_precall* ./processed/ &&
-				wait
-			fi
 			touch "${projdir}/call2_${samples_list}"
 
 			if [[ "$samples_list" == "samples_list_node_1.txt" ]] && test ! -f ${projdir}/snpcall/${pop}_${ref2%.f*}_${ploidy_ref2}x_raw.vcf*; then
@@ -2464,6 +2381,7 @@ main () {
 						fi
 						mv "${projdir}/snpcall/${i%.f*}_${ref3%.f*}.hold.g.vcf.gz" "${projdir}/snpcall/${i%.f*}_${ref3%.f*}.g.vcf.gz" && \
 						mv "${projdir}/snpcall/${i%.f*}_${ref3%.f*}.hold.g.vcf.gz.tbi" "${projdir}/snpcall/${i%.f*}_${ref3%.f*}.g.vcf.gz.tbi" &&
+						mv ${i%.f*}_${ref3%.f*}_precall.bam* ./processed/
 						wait
 					fi
 				fi ) &
@@ -2472,20 +2390,6 @@ main () {
 				fi
 			done
 			wait
-			if [[ $nodes -gt 1 ]] && test -f ${projdir}/GBSapp_run_node_1.sh; then
-				touch ${projdir}/queue_move_${samples_list%.txt}
-				queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-				while [[ "$queue_move" -gt 1 ]]; do
-					rm ${projdir}/queue_move_${samples_list%.txt}; sleep $[ ( $RANDOM % 120 )  + 30 ]s
-					touch ${projdir}/queue_move_${samples_list%.txt}
-					queue_move=$(ls ${projdir}/queue_move_samples_list_node_* | wc -l)
-				done
-				mv *_${ref3%.f*}_precall* ./processed/
-				rm ${projdir}/queue_move_${samples_list%.txt}
-			else
-				mv *_${ref3%.f*}_precall* ./processed/
-				wait
-			fi
 			touch "${projdir}/call3_${samples_list}"
 
 			if [[ "$samples_list" == "samples_list_node_1.txt" ]] && test ! -f ${projdir}/snpcall/${pop}_${ref3%.f*}_${ploidy_ref3}x_raw.vcf*; then
